requirements.txt
# Core Framework
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Telegram Bot
python-telegram-bot==21.0.1

# Authentication & Security
PyJWT==2.8.0
cryptography==41.0.7

# Task Queue
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# Cloud Storage
cloudinary==1.36.0

# Google Sheets API
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0

# QR Code Generation
qrcode[pil]==7.4.2
Pillow==10.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring
sentry-sdk==1.39.1